Moving CarPlayManagerDelegate into its own file, and documenting CarPlayManager initializer.
